apps/playground/requirements.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
pillow==10.1.0
plotly==5.17.0
python-dotenv==1.0.0
